Extract chat view contracts
